# docs/open/openapi.yml
openapi: "3.0.3"
info:
  title: "HTTP API"
  version: "6.0"
servers:
  - url: "http://localhost:5000/api/v6"
security:
  - jwt: []
  - ApiKeyAuth: []
...
paths:
...
  /movement:
    get:
      tags:
        - movement
      summary: List movements
      parameters:
        - name: movement_type  
          in: query
          schema:
            type: string
            enum: [MOVEMENT_INVENTORY, MOVEMENT_TOOL, MOVEMENT_PLAN, MOVEMENT_HEAD]
        - name: trans_code
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MovementResponse"
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - movement
      summary: Create a new movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/MovementPost"
      responses:
        "201":
          description: Movement created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movement"
        "401":
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /payment:
    get:
      tags:
        - payment
      summary: List payments
      parameters:
        - name: paid_date  
          in: query
          schema:
            type: string
            format: date
        - name: trans_code
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentResponse"
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - payment
      summary: Create a new payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/PaymentPost"
      responses:
        "201":
          description: Payment created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "401":
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /place:
    get:
      tags:
        - place
      summary: List places
      parameters:
        - name: place_type
          in: query
          schema:
            type: string
            enum: [PLACE_WAREHOUSE, PLACE_BANK, PLACE_CASH, PLACE_OTHER]
            description: Place type enumeration
        - name: place_name
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaceResponse"
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - place
      summary: Create a new place
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/PlacePost"
      responses:
        "201":
          description: Place created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
        "401":
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /project:
    get:
      tags:
        - project
      summary: List projects
      parameters:
        - name: project_name
          in: query
          schema:
            type: string
        - name: customer_code
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectResponse"
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - project
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/ProjectPost"
      responses:
        "201":
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  responses:
...
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MethodNotAllowed:
      description: Method not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    MetaTags:
      type: array
      items:
        type: string
      description: Additional tags for the data
      example: ["TAG1", "TAG2"]
    DataMap:
      type: object
      additionalProperties: true
      description: Flexible key-value map for additional metadata. The value is any json type.
      example:
        field_name: "Field value"
    Address:
      type: object
      properties:
        country:
          type: string
        state:
          type: string
        zip_code:
          type: string
        city:
          type: string
        street:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
          description: Additional tags for the data
          example: ["TAG1", "TAG2"]
        map:
          $ref: "#/components/schemas/DataMap"
    Contact:
      type: object
      properties:
        first_name:
          type: string
        surname:
          type: string
        status:
          type: string
        phone:
          type: string
        mobile:
          type: string
        email:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
          description: Additional tags for the data
          example: ["TAG1", "TAG2"]
        map:
          $ref: "#/components/schemas/DataMap"
    Event:
      type: object
      properties:
        uid:
          type: string
        subject:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        place:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
          description: Additional tags for the data
          example: ["TAG1", "TAG2"]
        map:
          $ref: "#/components/schemas/DataMap"
...
    Movement:
      type: object
      properties:
        movement_type:
          type: string
          enum: [MOVEMENT_INVENTORY, MOVEMENT_TOOL, MOVEMENT_PLAN, MOVEMENT_HEAD]
          description: ENUM field for movement type
        shipping_time:
          type: string
          format: date-time
          description: Date-time of shipping
        trans_code:
          type: string
          description: Reference to trans.code
        product_code:
          type: string
          description: Reference to Product.code
        tool_code:
          type: string
          description: Reference to Tool.code
        place_code:
          type: string
          description: Reference to Place.code
        movement_meta:
          type: object
          properties:
            qty:
              type: number
              format: float
            notes:
              type: string
            shared:
              type: boolean
            tags:
              $ref: "#/components/schemas/MetaTags"
        movement_map:
          $ref: "#/components/schemas/DataMap"
    MovementPost:
      type: object
      required:
        - movement_type
        - trans_code
        - shipping_time
      allOf:
        - properties:
            code:
              type: string
              description: Database independent unique external key. If not specified, it is generated on creation. It cannot be modified after creation.
              example: "MOV1731101982N123"
        - $ref: "#/components/schemas/Movement"
    MovementResponse:
      type: object
      allOf:
        - properties:
            id:
              type: integer
              format: int64
              description: Database primary key. Database dependent serial number type. Its value is unique only at table level and may change during data migrations.
        - $ref: "#/components/schemas/MovementPost"
        - properties:
            time_stamp:
              type: string
              format: date-time
              description: Timestamp of data creation
    Payment:
      type: object
      properties:
        paid_date:
          type: string
          format: date
          description: Date when payment was made
        trans_code:
          type: string
          description: Reference to trans.code
        payment_meta:
          type: object
          properties:
            amount:
              type: number
              format: double
            notes:
              type: string
            tags:
              $ref: "#/components/schemas/MetaTags"
        payment_map:
          $ref: "#/components/schemas/DataMap"
    PaymentPost:
      type: object
      required:
        - paid_date
        - trans_code
      allOf:
        - properties:
            code:
              type: string
              description: Database independent unique external key. If not specified, it is generated on creation. It cannot be modified after creation.
              example: "PMT1731101982N123"
        - $ref: "#/components/schemas/Payment"
    PaymentResponse:
      type: object
      allOf:
        - properties:
            id:
              type: integer
              format: int64
              description: Database primary key. Database dependent serial number type. Its value is unique only at table level and may change during data migrations.
        - $ref: "#/components/schemas/PaymentPost"
        - properties:
            time_stamp:
              type: string
              format: date-time
              description: Timestamp of data creation
    Place:
      type: object
      properties:
        place_type:
          type: string
          enum: [PLACE_WAREHOUSE, PLACE_BANK, PLACE_CASH, PLACE_OTHER]
          description: ENUM field for place type
        place_name:
          type: string
          description: The full name of the place.
        currency_code:
          type: string
          description: Reference to currency.code
        address:
          $ref: '#/components/schemas/Address'
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        place_meta:
          type: object
          properties:
            notes:
              type: string
            inactive:
              type: boolean
            tags:
              $ref: "#/components/schemas/MetaTags"
        place_map:
          $ref: "#/components/schemas/DataMap"
    PlacePost:
      type: object
      required:
        - place_type
        - place_name
      allOf:
        - properties:
            code:
              type: string
              description: Database independent unique external key. If not specified, it is generated on creation. It cannot be modified after creation.
              example: "PLC1731101982N123"
        - $ref: "#/components/schemas/Place"
    PlaceResponse:
      type: object
      allOf:
        - properties:
            id:
              type: integer
              format: int64
              description: Database primary key. Database dependent serial number type. Its value is unique only at table level and may change during data migrations.
        - $ref: "#/components/schemas/PlacePost"
        - properties:
            time_stamp:
              type: string
              format: date-time
              description: Timestamp of data creation
...
    Project:
      type: object
      properties:
        project_name:
          type: string
          description: The name of the project.
        customer_code:
          type: string
          description: Reference to Customer.code
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        contacts:
          type: array  
          items:
            $ref: '#/components/schemas/Contact'
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        project_meta:
          type: object
          properties:
            start_date:
              type: string
              format: date
            end_date:
              type: string
              format: date
            notes:
              type: string
            inactive:
              type: boolean
            tags:
              $ref: "#/components/schemas/MetaTags"
        project_map:
          $ref: "#/components/schemas/DataMap"
    ProjectPost:
      type: object
      required:
        - project_name
      allOf:
        - properties:
            code:
              type: string
              description: Database independent unique external key. If not specified, it is generated on creation. It cannot be modified after creation.
              example: "PRJ1731101982N123"
        - $ref: "#/components/schemas/Project"  
    ProjectResponse:  
      type: object
      allOf:
        - properties:
            id:
              type: integer
              format: int64
              description: Database primary key. Database dependent serial number type. Its value is unique only at table level and may change during data migrations.
        - $ref: "#/components/schemas/ProjectPost"
        - properties:
            time_stamp:
              type: string
              format: date-time
              description: Timestamp of data creation
...